requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
PyMySQL
python-dotenv
Werkzeug
Jinja2
